Move search components into src/search/
