+++ src/search/README.md
@@ -1,0 +1,3 @@
+Client implementation for elasticlunr.
+
+Main entry points is `LiveSearchBox`
