+++ src/search/index.js
@@ -1,0 +1,2 @@
+// Main entry point
+export { default as LiveSearchBox } from './components/LiveSearchBox'
